% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanUpNames.R
\name{cleanUpNames}
\alias{cleanUpNames}
\title{Clean Up Player Names For Merging.}
\usage{
cleanUpNames(name)
}
\arguments{
\item{name}{character vector of player names.}
}
\value{
Vector of cleaned player names.
}
\description{
Cleans up names of players for merging.
}
\details{
Cleans up names of NFL Football players, including making them all-caps,
removing common suffixes, punctuation, spaces, etc. This is helpful for
merging multiple datasets.
}
\examples{
oldNames <- c("Peyton Manning","Tom Brady","Marvin Harrison Jr.")
cleanNames <- cleanUpNames(oldNames)
cleanNames
}
\concept{fantasyFootball}
