% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.magic.R
\name{convert.magic}
\alias{convert.magic}
\title{Convert Variable Types.}
\usage{
convert.magic(obj, type)
}
\arguments{
\item{obj}{name of dataframe (object)}

\item{type}{type to convert variables to one of:
\itemize{
  \item \code{"character"}
  \item \code{"numeric"}
  \item \code{"factor"}
}}
}
\value{
Dataframe with columns converted to a particular type.
}
\description{
Converts variable types of multiple columns of a dataframe at once.
}
\details{
Converts variable types of multiple columns of a dataframe at once. Convert
variable types to character, numeric, or factor.
}
\examples{
# Prepare Data
data("USArrests")

# Convert variables to character
convert.magic(USArrests, "character")

}
\seealso{
\url{https://stackoverflow.com/questions/11261399/function-for-converting-dataframe-column-type/11263399#11263399}

Other dataManipulation:
\code{\link[=columnBindFill]{columnBindFill()}},
\code{\link[=dropColsWithAllNA]{dropColsWithAllNA()}},
\code{\link[=dropRowsWithAllNA]{dropRowsWithAllNA()}},
\code{\link[=varsDifferentTypes]{varsDifferentTypes()}}

Other conversion:
\code{\link[=convertHoursAMPM]{convertHoursAMPM()}},
\code{\link[=convertToHours]{convertToHours()}},
\code{\link[=convertToMinutes]{convertToMinutes()}},
\code{\link[=convertToSeconds]{convertToSeconds()}},
\code{\link[=percentileToTScore]{percentileToTScore()}},
\code{\link[=pom]{pom()}}
}
\concept{conversion}
\concept{dataManipulation}
