% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencyPerDuration.R
\name{timesPerInterval}
\alias{timesPerInterval}
\alias{computeItemFrequencies}
\title{Frequency Per Duration.}
\usage{
timesPerInterval(
  num_occurrences,
  interval,
  duration = "month",
  not_occurred_past_year
)

computeItemFrequencies(
  item_names,
  data,
  duration = "month",
  frequency_vars,
  interval_vars,
  not_in_past_year_vars
)
}
\arguments{
\item{num_occurrences}{The number of times the behavior occurred during the
specified interval, \code{interval}.}

\item{interval}{The specified interval corresponding to the number of times
the behavior occurred \code{num_occurrences}.}

\item{duration}{The desired duration during which to estimate how many times
the behavior occurred:
\itemize{
  \item \code{1} = average number of times per day
  \item \code{2} = average number of times per week
  \item \code{3} = number of times in the past month
  \item \code{4} = number of times in the past year
}}

\item{not_occurred_past_year}{Whether or not the behavior did NOT occur in
the past year. If \code{yes}, the behavior did not occur in the past year.}

\item{item_names}{The names of the questionnaire items.}

\item{data}{The data object.}

\item{frequency_vars}{The name(s) of the variables corresponding to the
number of occurrences (\code{num_occurrences}).}

\item{interval_vars}{The name(s) of the variables corresponding to the
intervals (\code{interval}).}

\item{not_in_past_year_vars}{The name(s) of the variables corresponding to
whether the behavior did not occur in the past year
(\code{not_occurred_past_year}).}
}
\value{
The frequency of the behavior for the specified duration.
}
\description{
Estimate frequency of a behavior for a particular duration.
}
\details{
Estimates the frequency of a given behavior for a particular duration, given
a specified number of times it occurred during a specified interval.
}
\examples{
timesPerInterval(
  num_occurrences = 2,
  interval = 3,
  duration = "month",
  not_occurred_past_year = 0
)
}
\concept{behaviorFrequency}
