% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationCombine.R
\name{imputationCombine}
\alias{imputationCombine}
\title{Combine Results from Mixed Effect Imputation Models.}
\usage{
imputationCombine(model, dig = 3)
}
\arguments{
\item{model}{name of \code{lme()} model object with multiply imputed data.}

\item{dig}{number of decimals to print in output.}
}
\value{
Summary of model fit and information for mixed effect imputation models.
}
\description{
Function that combines lme results across multiple imputation runs.
}
\details{
[INSERT].
}
\examples{
#INSERT
}
\seealso{
Other multipleImputation:
\code{\link[=imputationModelCompare]{imputationModelCompare()}},
\code{\link[=imputationPRV]{imputationPRV()}},
\code{\link[=lmCombine]{lmCombine()}}
}
\concept{multipleImputation}
