% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationModelCompare.R
\name{imputationModelCompare}
\alias{imputationModelCompare}
\title{Compare Mixed Effect Imputation Models.}
\usage{
imputationModelCompare(model1, model2)
}
\arguments{
\item{model1}{name of first \code{lme()} model object with multiply imputed
data.}

\item{model2}{name of second \code{lme()} model object with multiply imputed
data.}
}
\value{
Likelihood ratio test for model comparison of two mixed effect imputation
models.
}
\description{
Function that compares two nested \code{lme()} models from multiple
imputation using likelihood ratio test.
}
\details{
[INSERT].
}
\examples{
#INSERT
}
\seealso{
Other multipleImputation:
\code{\link[=imputationCombine]{imputationCombine()}},
\code{\link[=imputationPRV]{imputationPRV()}},
\code{\link[=lmCombine]{lmCombine()}}
}
\concept{multipleImputation}
