% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prvImputation.R
\name{imputationPRV}
\alias{imputationPRV}
\title{Proportional Reduction of Variance from Imputation Models.}
\usage{
imputationPRV(baseline, full, baselineTime = 1, fullTime = 1)
}
\arguments{
\item{baseline}{The baseline model object fit with the imputed data.}

\item{full}{The full model object fit with the imputed data.}

\item{baselineTime}{The position of the random effect of time (random slopes) among the random slopes in the baseline model. For example:
\itemize{
 \item \code{0} = no random slopes
 \item \code{1} = time is the 1st random effect
 \item \code{2} = time is the second random effect
}}

\item{fullTime}{The position of the random effect of time (random slopes) among the random slopes in the full model. For example:
\itemize{
 \item \code{0} = no random slopes
 \item \code{1} = time is the 1st random effect
 \item \code{2} = time is the second random effect
}}
}
\value{
The proportional reduction of variance from a baseline mixed-effects model
to a full mixed effects model.
}
\description{
Calculate the proportional reduction of variance in imputation models.
}
\details{
[INSERT].
}
\examples{
#INSERT
}
\seealso{
Other multipleImputation:
\code{\link[=imputationCombine]{imputationCombine()}},
\code{\link[=imputationModelCompare]{imputationModelCompare()}},
\code{\link[=lmCombine]{lmCombine()}}
}
\concept{multipleImputation}
