% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nan.R
\name{is.nan.data.frame}
\alias{is.nan.data.frame}
\title{NaN (Not a Number).}
\usage{
\method{is.nan}{data.frame}(x)
}
\arguments{
\item{x}{Dataframe.}
}
\value{
\code{TRUE} or \code{FALSE}, indicating whether values in a dataframe
are Not a Number (\code{NA}).
}
\description{
Check whether a value is "Not A Number" (\code{NaN}) in a dataframe.
}
\details{
[INSERT].
}
\examples{
# Prepare Data
df <- data.frame(item1 = rnorm(1000), item2 = rnorm(1000), item3 = rnorm(1000))
df[sample(1:nrow(df), size = 100), c("item1","item2","item3")] <- NaN

# Calculate Missingness-Adjusted Row Sum
is.nan(df)

}
\seealso{
\url{https://stackoverflow.com/questions/18142117/how-to-replace-nan-value-with-zero-in-a-huge-data-frame/18143097#18143097}

Other dataEvaluations:
\code{\link[=dropColsWithAllNA]{dropColsWithAllNA()}},
\code{\link[=dropRowsWithAllNA]{dropRowsWithAllNA()}},
\code{\link[=not_all_na]{not_all_na()}},
\code{\link[=not_any_na]{not_any_na()}}
}
\concept{dataEvaluations}
