% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isInstalled.R
\name{is_installed}
\alias{is_installed}
\title{Is Installed.}
\usage{
is_installed(mypkg)
}
\arguments{
\item{mypkg}{character vector of one or more packages}
}
\value{
\code{TRUE} or \code{FALSE}, indicating whether package is installed.
}
\description{
Checks whether a package is installed.
}
\details{
[INSERT].
}
\examples{
is_installed("nlme")
}
\seealso{
Other packages: 
\code{\link{getDependencies}()},
\code{\link{load_or_install}()}
}
\concept{packages}
