% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemInformationGRM.R
\name{calc_grm_probs}
\alias{calc_grm_probs}
\alias{itemInformationGRM}
\alias{get_thresholds}
\title{Item Information in graded response model in item response theory.}
\usage{
calc_grm_probs(a, b_thresholds, theta)

itemInformationGRM(a, b_thresholds, theta)

get_thresholds(x)
}
\arguments{
\item{a}{Discrimination parameter (slope).}

\item{b_thresholds}{Difficulty (severity) parameters (inflection point). Can
be provided either as a vector (for a given item) or as a list (for multiple
items).}

\item{theta}{Person's level on the construct.}

\item{x}{Dataframe.}
}
\value{
\code{calc_grm_probs()} returns the probability of each response category;
\code{itemInformationGRM()} returns the amount of item information.
}
\description{
Item information in graded response model in item response theory.
}
\details{
Estimates the amount of information provided by a given item in a graded
response model in item response theory as function of the item parameters
and the person's level on the construct (theta).
}
\examples{
calc_grm_probs(
 a = 1,
 b_thresholds = 0,
 theta = -4:4
)

calc_grm_probs(
 a = 1,
 b_thresholds = c(-1, 1),
 theta = -4:4
)

itemInformationGRM(
 a = 1,
 b_thresholds = 0,
 theta = -4:4
)

itemInformationGRM(
 a = 1,
 b_thresholds = c(-1, 1),
 theta = -4:4
)

itemParameters <- data.frame(
 item = c(1, 2, 3),
 a = c(0.5, 1, 1.5),
 b1 = c(-1, 0, 1),
 b2 = c(0, 1, 2),
 b3 = c(1, 2, 3)
)

calc_grm_probs(
 a = itemParameters$a,
 b_thresholds = get_thresholds(itemParameters),
 theta = -4:4)


itemInformationGRM(
 a = itemParameters$a,
 b_thresholds = get_thresholds(itemParameters),
 theta = -4:4)

}
\seealso{
Other IRT:
\code{\link[=deriv_d_negBinom]{deriv_d_negBinom()}},
\code{\link[=discriminationToFactorLoading]{discriminationToFactorLoading()}},
\code{\link[=fourPL]{fourPL()}},
\code{\link[=itemInformation]{itemInformation()}},
\code{\link[=reliabilityIRT]{reliabilityIRT()}},
\code{\link[=standardErrorIRT]{standardErrorIRT()}},
\code{\link[=test_info_4PL]{test_info_4PL()}}
}
\concept{IRT}
