% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmCombine.R
\name{lmCombine}
\alias{lmCombine}
\title{Combine Results from Multiple Regression Imputation Models.}
\usage{
lmCombine(model, dig = 3)
}
\arguments{
\item{model}{name of \code{lm()} model object with multiply imputed data.}

\item{dig}{number of decimals to print in output.}
}
\value{
Summary of multiple regression imputation models.
}
\description{
Function that combines \code{lm()} results across multiple imputation runs.
}
\details{
[INSERT].
}
\examples{
#INSERT
}
\seealso{
Other multipleImputation:
\code{\link[=imputationCombine]{imputationCombine()}},
\code{\link[=imputationModelCompare]{imputationModelCompare()}},
\code{\link[=imputationPRV]{imputationPRV()}}

Other multipleRegression:
\code{\link[=plot2WayInteraction]{plot2WayInteraction()}},
\code{\link[=ppPlot]{ppPlot()}},
\code{\link[=semPlotInteraction]{semPlotInteraction()}},
\code{\link[=update_nested]{update_nested()}}
}
\concept{multipleImputation}
\concept{multipleRegression}
