% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurementNoninvarianceEffectSize.R
\name{measurementNoninvarianceEffectSize}
\alias{measurementNoninvarianceEffectSize}
\title{Indexes of Effect Size for Measurement Noninvariance in Structural Equation
Models.}
\usage{
measurementNoninvarianceEffectSize(model_unconstrained, model_constrained)
}
\arguments{
\item{model_unconstrained}{\code{lavaan} model object for unconstrained
model.}

\item{model_constrained}{\code{lavaan} model object for constrained
model.}
}
\value{
\eqn{w} and \eqn{\Delta \mathrm{M}_{\mathrm{C}}}.
}
\description{
Function that estimates effect size indices for measurement noninvariance in
structural equation modeling.
}
\details{
From Newsom (2015):
Newsom, J. T. (2015). Longitudinal structural equation modeling: A
comprehensive introduction. Routledge.
}
\examples{
# Prepare Data
library("lavaan")

# Model
HS.model <- '
  visual =~ x1 + x2 + x3
            textual =~ x4 + x5 + x6
            speed   =~ x7 + x8 + x9
'

# configural invariance
fit1 <- lavaan::cfa(
  HS.model,
  data = HolzingerSwineford1939,
  group = "school")

# weak invariance
fit2 <- lavaan::cfa(
  HS.model,
  data = HolzingerSwineford1939,
  group = "school",
  group.equal = "loadings")

# strong invariance
fit3 <- lavaan::cfa(
  HS.model,
  data = HolzingerSwineford1939,
  group = "school",
  group.equal = c("intercepts", "loadings"))

# model comparison tests
lavaan::lavTestLRT(fit1, fit2, fit3)

# effect size of measurement noninvariance
measurementNoninvarianceEffectSize(
  model_unconstrained = fit1,
  model_constrained = fit2)

measurementNoninvarianceEffectSize(
  model_unconstrained = fit2,
  model_constrained = fit3)
}
\seealso{
Other structural equation modeling:
\code{\link[=equiv_chi]{equiv_chi()}},
\code{\link[=make_esem_model]{make_esem_model()}},
\code{\link[=puc]{puc()}},
\code{\link[=satorraBentlerScaledChiSquareDifferenceTestStatistic]{satorraBentlerScaledChiSquareDifferenceTestStatistic()}},
\code{\link[=semPlotInteraction]{semPlotInteraction()}}
}
\concept{structural equation modeling}
