% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortgage.R
\name{mortgage}
\alias{mortgage}
\title{Mortgage Principal and Interest.}
\usage{
mortgage(balance, interest, term = 30, n = 12)
}
\arguments{
\item{balance}{Initial mortgage balance.}

\item{interest}{Interest rate.}

\item{term}{Payoff period (in years).}

\item{n}{Number of payments per year.}
}
\value{
Amount of principal and interest payments.
}
\description{
Amount of principal and interest payments on a mortgage.
}
\details{
Calculates the amount of principal and interest payments on a mortgage.
}
\examples{
mortgage(balance = 300000, interest = .05)
mortgage(balance = 300000, interest = .04)
mortgage(balance = 300000, interest = .06)
mortgage(balance = 300000, interest = .05, term = 15)
}
