% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notAnyRowsNA.R
\name{not_any_na}
\alias{not_any_na}
\title{Not Any NA.}
\usage{
not_any_na(x)
}
\arguments{
\item{x}{column vector}
}
\value{
\code{TRUE} or \code{FALSE}, indicating whether the whole column does
not have any missing values (\code{NA}).
}
\description{
Check if all rows for a column are \code{NA}.
}
\details{
[INSERT].
}
\examples{
# Prepare Data
df <- data.frame(item1 = rnorm(1000), item2 = rnorm(1000), item3 = rnorm(1000))
df[sample(1:nrow(df), size = 100), "item2"] <- NA
df[,"item3"] <- NA

# Check if Not Any NA
not_any_na(df$item1)
not_any_na(df$item2)
not_any_na(df$item3)
}
\seealso{
Other dataEvaluations:
\code{\link[=dropColsWithAllNA]{dropColsWithAllNA()}},
\code{\link[=dropRowsWithAllNA]{dropRowsWithAllNA()}},
\code{\link[=is.nan.data.frame]{is.nan.data.frame()}},
\code{\link[=not_all_na]{not_all_na()}}
}
\concept{dataEvaluations}
