% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pValues.R
\name{pValue}
\alias{pValue}
\title{p-values.}
\usage{
pValue(value, digits = 3)
}
\arguments{
\item{value}{The p-value.}

\item{digits}{Number of decimal digits for printing the p-value.}
}
\value{
p-value.
}
\description{
Suppress the leading zero when printing p-values.
}
\details{
[INSERT].
}
\examples{
pValue(0.70)
pValue(0.04)
pValue(0.00002)
}
\seealso{
Other formatting:
\code{\link[=apa]{apa()}},
\code{\link[=specify_decimal]{specify_decimal()}},
\code{\link[=suppressLeadingZero]{suppressLeadingZero()}}
}
\concept{formatting}
