% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodeIntensity.R
\name{recode_intensity}
\alias{recode_intensity}
\alias{mark_intensity_as_zero}
\title{Recode Intensity.}
\usage{
recode_intensity(intensity, not_in_the_past_year = NULL, frequency = NULL)

mark_intensity_as_zero(
  item_names,
  data,
  not_in_past_year_vars = NULL,
  frequency_vars = NULL
)
}
\arguments{
\item{intensity}{The intensity of the behavior.}

\item{not_in_the_past_year}{Whether or not the behavior did NOT occur in
the past year. If \code{yes}, the behavior did not occur in the past year.}

\item{frequency}{The frequency of the behavior.}

\item{item_names}{The names of the questionnaire items.}

\item{data}{The data object.}

\item{not_in_past_year_vars}{The name(s) of the variables corresponding to
whether the behavior did not occur in the past year
(\code{not_occurred_past_year}).}

\item{frequency_vars}{The name(s) of the variables corresponding to the
number of occurrences (\code{num_occurrences}).}
}
\value{
The intensity of the behavior.
}
\description{
Recode intensity of behavior based on frequency of behavior.
}
\details{
Recodes the intensity of behavior to zero if the frequency of the behavior is
zero (i.e., if the behavior has not occurred).
}
\concept{behaviorIntensity}
