% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliabilityProductTerm.R
\name{reliabilityOfProductTerm}
\alias{reliabilityOfProductTerm}
\title{Reliability of Product Term.}
\usage{
reliabilityOfProductTerm(x, y, reliabilityX, reliabilityY)
}
\arguments{
\item{x}{Vector of one variable that is used in the computation of the
product term.}

\item{y}{Vector of second variable that is used in the computation of the
product term.}

\item{reliabilityX}{The reliability of the \code{x} variable.}

\item{reliabilityY}{The reliability of the \code{y} variable.}
}
\value{
Reliability of the product term that is computed from the multiplication of
\code{x} and \code{y}.
}
\description{
Estimate the reliability of a product term (assuming the individual indices
are centered for the computation of the product term).
}
\details{
Estimates the reliability of a product term (assuming the individual indices
are centered for the computation of the product term).
}
\examples{
v1 <- rnorm(1000, mean = 100, sd = 15)
v2 <- rnorm(1000, mean = 1, sd = 15)
reliabilityOfProductTerm(x = v1, y = v2,
 reliabilityX = .7, reliabilityY = .8)
}
\seealso{
Other reliability:
\code{\link[=reliabilityOfDifferenceScore]{reliabilityOfDifferenceScore()}},
\code{\link[=repeatability]{repeatability()}}
}
\concept{reliability}
