% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setLabPath.R
\name{setLabPath}
\alias{setLabPath}
\title{Set Lab Path.}
\usage{
setLabPath()
}
\value{
The object \code{petersenLab} with containing the path directory to the lab
drive.
}
\description{
Sets the path directory to the lab drive.
}
\details{
Sets the path directory to the lab drive, and saves it in the object
\code{petersenLab}.
}
\examples{
\dontrun{
petersenLabPath <- setLabPath()
}
}
\concept{lab}
