% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageDependencies.R
\name{getDependencies}
\alias{getDependencies}
\title{Package Dependencies.}
\usage{
getDependencies(packs)
}
\arguments{
\item{packs}{Character vector of names of target packages.}
}
\value{
Vector of packages that depend on the target package(s).
}
\description{
Determine package dependencies.
}
\details{
Determine which packages depend on a target package (or packages).
}
\examples{
options(repos = "https://cran.r-project.org")
getDependencies("tidyverse")

}
\seealso{
\url{https://stackoverflow.com/questions/52929114/install-packages-in-r-without-internet-connection-with-all-dependencies/52935020#52935020}

Other packages: 
\code{\link{is_installed}()},
\code{\link{load_or_install}()}
}
\concept{packages}
