% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{Mode}
\alias{Mode}
\title{Statistical Mode.}
\usage{
Mode(x, multipleModes = "all")
}
\arguments{
\item{x}{Numerical vector.}

\item{multipleModes}{How to handle multiple modes. One of:
\itemize{
  \item \code{"mean"} = if there are multiple modes, take the mean of all
  modes
  \item \code{"first"} = if there are multiple modes, select the first mode
  \item \code{"all"} = if there are multiple modes, return all modes
}}
}
\value{
Statistical mode(s).
}
\description{
Calculate statistical mode.
}
\details{
Calculates statistical mode(s).
}
\examples{
# Prepare Data
v1 <- c(1, 1, 2, 2, 3)

#Calculate Statistical Mode
Mode(v1)
Mode(v1, multipleModes = "mean")
Mode(v1, multipleModes = "first")

}
\seealso{
\url{https://stackoverflow.com/questions/2547402/how-to-find-the-statistical-mode/8189441#8189441}

Other computations: 
\code{\link{meanSum}()},
\code{\link{mySum}()}
}
\concept{computations}
