% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomogrammer.R
\name{nomogrammer}
\alias{nomogrammer}
\title{Create Nomogram.}
\usage{
nomogrammer(
  Prevalence,
  Sens = NULL,
  Spec = NULL,
  Plr = NULL,
  Nlr = NULL,
  Detail = FALSE,
  NullLine = FALSE,
  LabelSize = (14/5),
  Verbose = FALSE
)
}
\arguments{
\item{Prevalence}{Prevalence (base rate/prior probability) of characteristic,
as a number between 0 and 1.}

\item{Sens}{Sensitivity of the test at a given cut point, as a number
between 0 and 1.}

\item{Spec}{Specificity of the test at a given cut point, as a number
between 0 and 1.}

\item{Plr}{Positive likelihood ratio of the test at a given cut point.}

\item{Nlr}{Positive likelihood ratio of the test at a given cut point.}

\item{Detail}{If \code{TRUE}, overlay key statistics onto the plot.}

\item{NullLine}{If \code{TRUE}, add a line from prior prob through LR = 1.}

\item{LabelSize}{Label size.}

\item{Verbose}{Print out relevant metrics in the console.}
}
\value{
ggplot object.
}
\description{
Create nomogram plot.
}
\details{
Create nomogram plot from the following at a given cut point:

\itemize{
  \item 1) base rate, and either
  \item 2) sensitivity and specificity, OR
  \item 3) positive likelihood ratio and negative likelihood ratio
}
}
\examples{
nomogrammer(
  Prevalence = .60,
  Plr = 12,
  Nlr = 0.6)

nomogrammer(
  Prevalence = .60,
  Sens = 0.421,
  Spec = 0.965)
}
