% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anyRowsNA.R
\name{not_all_na}
\alias{not_all_na}
\title{Any Rows Not NA.}
\usage{
not_all_na(x)
}
\arguments{
\item{x}{vector or column}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Check if any rows for a column are not \code{NA}.
}
\details{
Determine whether any rows for a column (or vector) are not missing
(\code{NA}).
}
\examples{
# Prepare Data
data("USArrests")

# Check if any rows are not NA
not_all_na(USArrests$Murder)
}
\seealso{
Other dataEvaluations: 
\code{\link{dropColsWithAllNA}()},
\code{\link{dropRowsWithAllNA}()},
\code{\link{is.nan.data.frame}()},
\code{\link{not_any_na}()}
}
\concept{dataEvaluations}
