% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentileToTscore.R
\name{percentileToTScore}
\alias{percentileToTScore}
\title{Percentile to T-Score Conversion.}
\usage{
percentileToTScore(percentileRank)
}
\arguments{
\item{percentileRank}{Vector of percentile ranks.}
}
\value{
Vector of T-scores.
}
\description{
Conversion of percentile ranks to T-scores.
}
\details{
Converts percentile ranks to the equivalent T-scores.
}
\examples{
percentileRanks <- c(1, 10, 20, 30, 40, 50, 60, 70, 80, 90, 99)

percentileToTScore(percentileRanks)
}
\seealso{
Other conversion: 
\code{\link{convert.magic}()},
\code{\link{convertHoursAMPM}()},
\code{\link{convertToHours}()},
\code{\link{convertToMinutes}()},
\code{\link{convertToSeconds}()},
\code{\link{pom}()}
}
\concept{conversion}
