% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEncryptedData.R
\name{read.aes}
\alias{read.aes}
\title{Read Encrypted Data.}
\usage{
read.aes(filename, key)
}
\arguments{
\item{filename}{Location of encrypted data.}

\item{key}{Encryption key.}
}
\value{
Unencrypted data.
}
\description{
Read data from encrypted file.
}
\details{
Reads data from an encrypted file. To write an data to an encrypted file, see
\link{write.aes}.
}
\examples{
# Location of Encryption Key on Local Computer (where only you should have access to it)
encryptionKeyLocation <- file.path(getwd(), "/encryptionKey.RData",
  fsep = "") #Can change to a different path, e.g.: "C:/Users/[USERNAME]/"

# Generate Encryption Key
key <- as.raw(sample(1:16, 16))

# Save Encryption Key
save(key, file = encryptionKeyLocation)

# Specify Credentials
credentials <- "Insert My Credentials Here"

# Save Encrypted Credentials
write.aes(df = credentials,
  filename = file.path(getwd(), "/encrypytedCredentials.txt", fsep = ""),
  key = key) #Change the file location to save this on the lab drive
rm(credentials)
rm(key)

# Read and Unencrypt the Credentials Using the Encryption Key
load(encryptionKeyLocation)
credentials <- read.aes(filename =
  file.path(getwd(), "/encrypytedCredentials.txt", fsep = ""), key = key)

}
\seealso{
\url{https://stackoverflow.com/questions/25318800/how-do-i-read-an-encrypted-file-from-disk-with-r/25321586#25321586}

Other encrypted: 
\code{\link{write.aes}()}
}
\concept{encrypted}
