% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specifyDecimals.R
\name{specify_decimal}
\alias{specify_decimal}
\title{Specify Decimals.}
\usage{
specify_decimal(x, k)
}
\arguments{
\item{x}{Numeric vector.}

\item{k}{Number of decimals to print.}
}
\value{
Character vector of numbers with the specified number of decimal
places.
}
\description{
Specify the number of decimals to print.
}
\details{
[INSERT].
}
\examples{
# Prepare Data
v1 <- rnorm(1000)

# Specify Decimals
specify_decimal(v1, 2)
}
