% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppressLeadingZero.R
\name{suppressLeadingZero}
\alias{suppressLeadingZero}
\title{Suppress Leading Zero.}
\usage{
suppressLeadingZero(value)
}
\arguments{
\item{value}{Numeric vector.}
}
\value{
Character vector of numbers without leading zeros.
}
\description{
Suppress leading zero of numbers.
}
\details{
[INSERT].
}
\examples{
# Prepare Data
v1 <- rnorm(1000)

# Suppress Leading Zero
suppressLeadingZero(v1)
}
\seealso{
Other formatting: 
\code{\link{apa}()},
\code{\link{pValue}()},
\code{\link{specify_decimal}()}
}
\concept{formatting}
